+++ android/app/src/main/java/com/onair/hearit/presentation/HearitClickListener.kt
@@ -1,4 +1,4 @@
-package com.onair.hearit.presentation.home
+package com.onair.hearit.presentation
 
 interface HearitClickListener {
     fun onClick(hearitId: Long)
