+++ android/app/src/main/java/com/onair/hearit/presentation/search/result/SearchResultFragment.kt
@@ -16,9 +16,9 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.onair.hearit.databinding.FragmentSearchResultBinding
 import com.onair.hearit.domain.model.SearchInput
+import com.onair.hearit.presentation.HearitClickListener
 import com.onair.hearit.presentation.MainActivity
 import com.onair.hearit.presentation.detail.PlayerDetailActivity
-import com.onair.hearit.presentation.home.HearitClickListener
 
 class SearchResultFragment :
     Fragment(),
@@ -55,10 +55,6 @@ class SearchResultFragment :
         setupWindowInsets()
         setupRecyclerView()
         observeViewModel()
-        binding.nsvSearchResult.setOnTouchListener { _, _ ->
-            hideKeyboard()
-            false
-        }
     }
 
     private fun setupWindowInsets() {
