+++ android/app/src/main/res/layout/item_category_hearit.xml
@@ -15,7 +15,7 @@
 
         <variable
             name="hearitClickListener"
-            type="com.onair.hearit.presentation.home.HearitClickListener" />
+            type="com.onair.hearit.presentation.HearitClickListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
