+++ android/app/src/main/res/layout/item_searched_hearit.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="clickListener"
-            type="com.onair.hearit.presentation.home.HearitClickListener" />
+            type="com.onair.hearit.presentation.HearitClickListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
